AppDa

+++ app/src/main/java/com/js/nowakelock/data/db/entity/AppCount.kt
@@ -15,6 +15,6 @@ data class AppCount(
     var packageName: String,
     var count: Int = 0,
     var countTime: Long = 0,
-    var blockCount: Long = 0,
+    var blockCount: Int = 0,
     var blockCountTime: Long = 0
 )
